Added panel inside panel shrink.

+++ src/com/ui/GuiElement.java
@@ -20,9 +20,9 @@ public abstract class GuiElement {
     
     // ---- GETTERS & SETTERS ----
     public Rectangle getBounds() { return new Rectangle(this.x, this.y, this.w, this.h); }
-    public float getX() { return x; }
+    public int getX() { return x; }
     public void setX(int x) { this.x = x; }
-    public float getY() { return y; }
+    public int getY() { return y; }
     public void setY(int y) { this.y = y; }
     public boolean isEnabled() { return isEnabled; }
     public void setEnabled(boolean isEnabled) { this.isEnabled = isEnabled; }
